Update Entity/Model folder naming for consistency
