+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/assemblers/ForestAreaCodeResourceAssembler.java
@@ -7,10 +7,10 @@ import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.server.mvc.RepresentationModelAssemblerSupport;
 import org.springframework.stereotype.Component;
 
-import ca.bc.gov.nrs.wfprev.data.model.ForestAreaCodeEntity;
 import ca.bc.gov.nrs.wfprev.common.enums.CodeTables;
 import ca.bc.gov.nrs.wfprev.controllers.CodesController;
-import ca.bc.gov.nrs.wfprev.data.resources.ForestAreaCodeModel;
+import ca.bc.gov.nrs.wfprev.data.entities.ForestAreaCodeEntity;
+import ca.bc.gov.nrs.wfprev.data.models.ForestAreaCodeModel;
 
 @Component
 public class ForestAreaCodeResourceAssembler extends RepresentationModelAssemblerSupport<ForestAreaCodeEntity, ForestAreaCodeModel> {
