+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/entities/ProgramAreaEntity.java
@@ -1,8 +1,9 @@
-package ca.bc.gov.nrs.wfprev.data.model;
+package ca.bc.gov.nrs.wfprev.data.entities;
 
 import java.io.Serializable;
-
 import java.util.Date;
+
+import org.hibernate.annotations.UuidGenerator;
 import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedBy;
@@ -10,9 +11,11 @@ import org.springframework.data.annotation.LastModifiedDate;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-import jakarta.persistence.Id;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 import jakarta.validation.constraints.NotNull;
@@ -23,35 +26,23 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "general_scope_code")
+@Table(name = "program_area")
 @JsonIgnoreProperties(ignoreUnknown = false)
 @Data
 @EqualsAndHashCode(callSuper = false)
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class GeneralScopeCodeEntity implements Serializable {
-  private static final long serialVersionUID = 1L;
-  
+public class ProgramAreaEntity implements Serializable {
   @Id
-  @Column(name = "general_scope_code")
-  @NotNull
-  private String generalScopeCode;
+  @UuidGenerator
+  @GeneratedValue(strategy = GenerationType.UUID)
+  @Column(name = "program_area_guid")
+  private String programAreaGuid;
 
   @NotNull
-  @Column(name = "description", length = 200)
-  private String description;
-
-  @Column(name = "display_order", length = 3)
-  private Integer displayOrder;
-
-  @NotNull
-  @Column(name = "effective_date", length = 200)
-  private Date effectiveDate;
-
-  @NotNull
-  @Column(name = "expiry_date", length = 200)
-  private Date expiryDate;
+	@Column(name="program_area_name", length = 100)
+	private String programAreaName;
 
   @Column(name = "revision_count", columnDefinition="Decimal(10) default '0'")
   @NotNull
