+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/models/ProgramAreaModel.java
@@ -1,12 +1,10 @@
-package ca.bc.gov.nrs.wfprev.data.resources;
-
-import java.util.Date;
+package ca.bc.gov.nrs.wfprev.data.models;
 
 import org.springframework.hateoas.server.core.Relation;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonRootName;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
 import ca.bc.gov.nrs.wfprev.common.entities.CommonModel;
 import lombok.AllArgsConstructor;
@@ -17,16 +15,13 @@ import lombok.NoArgsConstructor;
 
 @Data
 @EqualsAndHashCode(callSuper = false)
-@JsonRootName(value = "projectTypeCode")
-@Relation(collectionRelation = "projectTypeCode")
+@JsonRootName(value = "programArea")
+@Relation(collectionRelation = "programArea")
 @JsonInclude(Include.NON_NULL)
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProjectTypeCodeModel extends CommonModel<ProjectTypeCodeModel> {
-  private String projectTypeCode;
-  private String description;
-	private Integer displayOrder;
-	private Date effectiveDate;
-	private Date expiryDate;
+public class ProgramAreaModel extends CommonModel<ProgramAreaModel> {
+  private String programAreaGuid;
+	private String programAreaName;
 }
