+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/models/ProjectTypeCodeModel.java
@@ -1,10 +1,12 @@
-package ca.bc.gov.nrs.wfprev.data.resources;
+package ca.bc.gov.nrs.wfprev.data.models;
+
+import java.util.Date;
 
 import org.springframework.hateoas.server.core.Relation;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonRootName;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonRootName;
 
 import ca.bc.gov.nrs.wfprev.common.entities.CommonModel;
 import lombok.AllArgsConstructor;
@@ -15,13 +17,16 @@ import lombok.NoArgsConstructor;
 
 @Data
 @EqualsAndHashCode(callSuper = false)
-@JsonRootName(value = "programArea")
-@Relation(collectionRelation = "programArea")
+@JsonRootName(value = "projectTypeCode")
+@Relation(collectionRelation = "projectTypeCode")
 @JsonInclude(Include.NON_NULL)
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ProgramAreaModel extends CommonModel<ProgramAreaModel> {
-  private String programAreaGuid;
-	private String programAreaName;
+public class ProjectTypeCodeModel extends CommonModel<ProjectTypeCodeModel> {
+  private String projectTypeCode;
+  private String description;
+	private Integer displayOrder;
+	private Date effectiveDate;
+	private Date expiryDate;
 }
