+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/repositories/ForestAreaCodeRepository.java
@@ -3,7 +3,7 @@ package ca.bc.gov.nrs.wfprev.data.repositories;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import ca.bc.gov.nrs.wfprev.common.repository.CommonRepository;
-import ca.bc.gov.nrs.wfprev.data.model.ForestAreaCodeEntity;
+import ca.bc.gov.nrs.wfprev.data.entities.ForestAreaCodeEntity;
 
 @RepositoryRestResource(exported = false)
 public interface ForestAreaCodeRepository extends CommonRepository<ForestAreaCodeEntity, String> {
