+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/repositories/GeneralScopeCodeRepository.java
@@ -1,8 +1,8 @@
 package ca.bc.gov.nrs.wfprev.data.repositories;
 
-import ca.bc.gov.nrs.wfprev.data.model.GeneralScopeCodeEntity;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import ca.bc.gov.nrs.wfprev.common.repository.CommonRepository;
+import ca.bc.gov.nrs.wfprev.data.entities.GeneralScopeCodeEntity;
 
 @RepositoryRestResource(exported = false)
 public interface GeneralScopeCodeRepository extends CommonRepository<GeneralScopeCodeEntity, String> {
