+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/repositories/ProgramAreaRepository.java
@@ -1,8 +1,8 @@
 package ca.bc.gov.nrs.wfprev.data.repositories;
 
-import ca.bc.gov.nrs.wfprev.data.model.ProgramAreaEntity;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import ca.bc.gov.nrs.wfprev.common.repository.CommonRepository;
+import ca.bc.gov.nrs.wfprev.data.entities.ProgramAreaEntity;
 
 @RepositoryRestResource(exported = false)
 public interface ProgramAreaRepository extends CommonRepository<ProgramAreaEntity, String> {
