+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/repositories/ProjectBoundaryRepository.java
@@ -1,8 +1,8 @@
 package ca.bc.gov.nrs.wfprev.data.repositories;
 
-import ca.bc.gov.nrs.wfprev.data.model.ProjectBoundaryEntity;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import ca.bc.gov.nrs.wfprev.common.repository.CommonRepository;
+import ca.bc.gov.nrs.wfprev.data.entities.ProjectBoundaryEntity;
 
 @RepositoryRestResource(exported = false)
 public interface ProjectBoundaryRepository extends CommonRepository<ProjectBoundaryEntity, String> {
