+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/data/repositories/ProjectRepository.java
@@ -1,8 +1,9 @@
 package ca.bc.gov.nrs.wfprev.data.repositories;
 
 import ca.bc.gov.nrs.wfprev.common.repository.CommonRepository;
+import ca.bc.gov.nrs.wfprev.data.entities.ProjectEntity;
+
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import ca.bc.gov.nrs.wfprev.data.model.ProjectEntity;
 
 @RepositoryRestResource(exported = false)
 public interface ProjectRepository extends CommonRepository<ProjectEntity, String> {
