+++ server/wfprev-api/src/main/java/ca/bc/gov/nrs/wfprev/services/ProjectService.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Component;
 import ca.bc.gov.nrs.wfone.common.service.api.ServiceException;
 import ca.bc.gov.nrs.wfprev.common.services.CommonService;
 import ca.bc.gov.nrs.wfprev.data.assemblers.ProjectResourceAssembler;
-import ca.bc.gov.nrs.wfprev.data.model.ProjectEntity;
+import ca.bc.gov.nrs.wfprev.data.entities.ProjectEntity;
+import ca.bc.gov.nrs.wfprev.data.models.ProjectModel;
 import ca.bc.gov.nrs.wfprev.data.repositories.ProjectRepository;
-import ca.bc.gov.nrs.wfprev.data.resources.ProjectModel;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
